rollingCookies override destroySession (#126)
when you use `destroySession` the `Set-Cookie` header is assigned an
empty string. actually the `rollingCookies` function checks if the
header value is Falsy, and if it is it takes the session from the
request and commits it which makes it impossible to logout.

The function should check if the value is strictly null instead of Falsy
to avoid overwriting the `Set-Cookie` header when trying to logout.