change exports behavior slightly (#224)
This follows what your original thinking was and it simplifies things
because TypeScript is able to pick up the types for these files without
any issues.

I personally prefer this because it seems to do a better job of grouping
similar utilities within a "subpackage" concept. But whatever you want
to do I'm fine with.