Fix types for using .then() with promiseHash (#458)
Using `.then()` in a value passed to `promiseHash()` produces
`Promise<unknown>`. This should fix that up so that it properly infers
the type.

I'm going to be 100% honest, I'm not entirely sure as to why this works
and the existing type constraint _doesn't_. But it seems to do what I
expect.

It would be awesome if this could also get integrated into the v7 line
for use with `remix@2` as well 🙏