Expose promise from `useDebouncedSubmit` to align with `useSubmit` (#477)
Currently `useDebounceSubmit` has a bit of a [Zalg̯o
issuḙ͖](https://blog.izs.me/2013/08/designing-apis-for-asynchrony/)
issue, in that it sometimes returns a `Promise` and sometimes doesn't.

This incidentally makes the return type of the returned `submit`
function `Promise<void> | void`, which causes issues when attempting to
pass the submit handler to a prop that's expecting something compatible
with the function returned by `useSubmit`.

Instead, it should return a Promise that resolves when the debounced
`submit()` function also resolves.